Add graphql function builder
