+++ service-model-generator/modules/service-model-generator-ls-extension/src/main/java/io/ballerina/servicemodelgenerator/extension/function/HttpFunctionBuilder.java
@@ -20,17 +20,33 @@ package io.ballerina.servicemodelgenerator.extension.function;
 
 import com.google.gson.Gson;
 import com.google.gson.stream.JsonReader;
+import io.ballerina.compiler.api.SemanticModel;
+import io.ballerina.compiler.syntax.tree.FunctionDefinitionNode;
+import io.ballerina.modelgenerator.commons.Annotation;
+import io.ballerina.modelgenerator.commons.ServiceDatabaseManager;
 import io.ballerina.servicemodelgenerator.extension.model.Function;
 import io.ballerina.servicemodelgenerator.extension.model.GetModelContext;
+import io.ballerina.servicemodelgenerator.extension.model.ModelFromSourceContext;
+import io.ballerina.servicemodelgenerator.extension.model.Value;
 import io.ballerina.servicemodelgenerator.extension.util.Utils;
 
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.nio.charset.StandardCharsets;
+import java.util.List;
+import java.util.Map;
 import java.util.Optional;
+import java.util.Set;
+import java.util.stream.Collectors;
 
+import static io.ballerina.servicemodelgenerator.extension.ServiceModelGeneratorConstants.KIND_RESOURCE;
+import static io.ballerina.servicemodelgenerator.extension.util.Constants.BALLERINA;
 import static io.ballerina.servicemodelgenerator.extension.util.Constants.HTTP;
+import static io.ballerina.servicemodelgenerator.extension.util.Constants.OBJECT_METHOD;
+import static io.ballerina.servicemodelgenerator.extension.util.Constants.RESOURCE;
+import static io.ballerina.servicemodelgenerator.extension.util.Utils.getFunctionModel;
+import static io.ballerina.servicemodelgenerator.extension.util.Utils.updateValue;
 
 public class HttpFunctionBuilder extends AbstractFunctionBuilder {
     private static final String HTTP_FUNCTION_MODEL_LOCATION = "functions/http_%s.json";
@@ -51,7 +67,68 @@ public class HttpFunctionBuilder extends AbstractFunctionBuilder {
     }
 
     @Override
+    public Function getModelFromSource(ModelFromSourceContext context) {
+        return getFunctionFromFunctionDef((FunctionDefinitionNode) context.node(), context.semanticModel());
+    }
+
+    @Override
     public String kind() {
         return HTTP;
+    }
+
+    public static Function getFunctionFromFunctionDef(FunctionDefinitionNode functionDefinitionNode,
+                                                      SemanticModel semanticModel) {
+        ServiceDatabaseManager databaseManager = ServiceDatabaseManager.getInstance();
+        List<Annotation> annotationAttachments = databaseManager.
+                getAnnotationAttachments(BALLERINA, HTTP, OBJECT_METHOD);
+        Map<String, Value> annotations = Function.createAnnotationsMap(annotationAttachments);
+        Function functionModel = getFunctionModel(functionDefinitionNode, semanticModel, true, false, annotations);
+        functionModel.setEditable(true);
+
+        if (functionModel.getKind().equals(KIND_RESOURCE)) {
+            Optional<Function> resourceFunctionOp = getHttpFunctionModel(RESOURCE);
+            if (resourceFunctionOp.isPresent()) {
+                Function resourceFunction = resourceFunctionOp.get();
+                if (resourceFunction.getReturnType().getResponses().size() > 1) {
+                    resourceFunction.getReturnType().getResponses().remove(1);
+                }
+                updateFunctionInfo(resourceFunction, functionModel);
+                return resourceFunction;
+            }
+        } else {
+            functionModel.setAnnotations(null);
+            functionModel.getAccessor().setEnabled(false);
+        }
+        return functionModel;
+    }
+
+    private static Optional<Function> getHttpFunctionModel(String functionType) {
+        InputStream resourceStream = Utils.class.getClassLoader()
+                .getResourceAsStream(String.format(HTTP_FUNCTION_MODEL_LOCATION, functionType));
+        if (resourceStream == null) {
+            return Optional.empty();
+        }
+
+        try (JsonReader reader = new JsonReader(new InputStreamReader(resourceStream, StandardCharsets.UTF_8))) {
+            return Optional.of(new Gson().fromJson(reader, Function.class));
+        } catch (IOException e) {
+            return Optional.empty();
+        }
+    }
+
+    private static void updateFunctionInfo(Function functionModel, Function commonFunction) {
+        functionModel.setEditable(commonFunction.isEditable());
+        functionModel.setEnabled(true);
+        functionModel.setKind(commonFunction.getKind());
+        functionModel.setCodedata(commonFunction.getCodedata());
+        updateValue(functionModel.getAccessor(), commonFunction.getAccessor());
+        updateValue(functionModel.getName(), commonFunction.getName());
+        updateValue(functionModel.getReturnType(), commonFunction.getReturnType());
+        Set<String> existingTypes = functionModel.getParameters().stream()
+                .map(parameter -> parameter.getType().getValue())
+                .collect(Collectors.toSet());
+        commonFunction.getParameters().stream()
+                .filter(commonParam -> !existingTypes.contains(commonParam.getType().getValue()))
+                .forEach(functionModel::addParameter);
     }
 }
