+++ service-model-generator/modules/service-model-generator-ls-extension/src/main/java/io/ballerina/servicemodelgenerator/extension/util/Constants.java
@@ -10,6 +10,13 @@ public final class Constants {
     public static final String HTTP = "http";
     public static final String AI = "ai";
     public static final String TCP = "tcp";
+    public static final String GRAPHQL = "graphql";
+
+    // annotation attachment points
+    public static final String OBJECT_METHOD = "OBJECT_METHOD";
+
+    // function types
+    public static final String RESOURCE = "resource";
 
     public static final String DEFAULT = "default";
